positional_encoding -> attention, incorporate copilot suggestions
