--- kale/embed/cnn.py
@@ -1,80 +1,0 @@
-"""
-DeepDTA based models for a drug-target interaction prediction problem.
-"""
-
-import torch.nn as nn
-import torch.nn.functional as F
-
-
-class CNNEncoder(nn.Module):
-    r"""
-    The DeepDTA's CNN encoder module, which comprises three 1D-convolutional layers and one max-pooling layer.
-    The module is applied to encoding drug/target sequence information, and the input should be transformed information
-    with integer/label encoding. The original paper is `"DeepDTA: deep drug–target binding affinity prediction"
-    <https://academic.oup.com/bioinformatics/article/34/17/i821/5093245>`_ .
-
-    Args:
-        num_embeddings (int): Number of embedding labels/categories, depends on the types of encoding sequence.
-        embedding_dim (int): Dimension of embedding labels/categories.
-        sequence_length (int): Max length of an input sequence.
-        num_kernels (int): Number of kernels (filters).
-        kernel_length (int): Length of kernel (filter).
-    """
-
-    def __init__(self, num_embeddings, embedding_dim, sequence_length, num_kernels, kernel_length):
-        super(CNNEncoder, self).__init__()
-        self.embedding = nn.Embedding(num_embeddings + 1, embedding_dim)
-        self.conv1 = nn.Conv1d(in_channels=sequence_length, out_channels=num_kernels, kernel_size=kernel_length)
-        self.conv2 = nn.Conv1d(in_channels=num_kernels, out_channels=num_kernels * 2, kernel_size=kernel_length)
-        self.conv3 = nn.Conv1d(in_channels=num_kernels * 2, out_channels=num_kernels * 3, kernel_size=kernel_length)
-        self.global_max_pool = nn.AdaptiveMaxPool1d(output_size=1)
-
-    def forward(self, x):
-        x = self.embedding(x)
-        x = F.relu(self.conv1(x))
-        x = F.relu(self.conv2(x))
-        x = F.relu(self.conv3(x))
-        x = self.global_max_pool(x)
-        x = x.squeeze(2)
-        return x
-
-
-class ProteinCNN(nn.Module):
-    """
-    A protein feature extractor using Convolutional Neural Networks (CNNs).
-
-    This class extracts features from protein sequences using a series of 1D convolutional layers.
-    The input protein sequence is first embedded and then passed through multiple convolutional
-    and batch normalization layers to produce a fixed-size feature vector.
-
-    Args:
-        embedding_dim (int): Dimensionality of the embedding space for protein sequences.
-        num_filters (list of int): A list specifying the number of filters for each convolutional layer.
-        kernel_size (list of int): A list specifying the kernel size for each convolutional layer.
-        padding (bool): Whether to apply padding to the embedding layer.
-    """
-
-    def __init__(self, embedding_dim, num_filters, kernel_size, padding=True):
-        super(ProteinCNN, self).__init__()
-        if padding:
-            self.embedding = nn.Embedding(26, embedding_dim, padding_idx=0)
-        else:
-            self.embedding = nn.Embedding(26, embedding_dim)
-        in_ch = [embedding_dim] + num_filters
-        # self.in_ch = in_ch[-1]
-        kernels = kernel_size
-        self.conv1 = nn.Conv1d(in_channels=in_ch[0], out_channels=in_ch[1], kernel_size=kernels[0])
-        self.bn1 = nn.BatchNorm1d(in_ch[1])
-        self.conv2 = nn.Conv1d(in_channels=in_ch[1], out_channels=in_ch[2], kernel_size=kernels[1])
-        self.bn2 = nn.BatchNorm1d(in_ch[2])
-        self.conv3 = nn.Conv1d(in_channels=in_ch[2], out_channels=in_ch[3], kernel_size=kernels[2])
-        self.bn3 = nn.BatchNorm1d(in_ch[3])
-
-    def forward(self, v):
-        v = self.embedding(v.long())
-        v = v.transpose(2, 1)
-        v = self.bn1(F.relu(self.conv1(v)))
-        v = self.bn2(F.relu(self.conv2(v)))
-        v = self.bn3(F.relu(self.conv3(v)))
-        v = v.view(v.size(0), v.size(2), -1)
-        return v
