+++ tests/embed/test_cnn.py
@@ -1,6 +1,6 @@
 import torch
 
-from kale.embed.cnn import CNNEncoder, ProteinCNN
+from kale.embed.attention_cnn import CNNEncoder, ProteinCNN
 
 
 def test_cnn_encoder():
